requirements.txt
google-cloud-aiplatform==1.69.0
tenacity==9.0.0
Flask==3.0.3
Flask-Cors==5.0.0
